・リザルトの俯瞰処理を追加 ・シーン遷移クラスの作成

+++ CaseStudy/CaseStudy/res/cpp/System/MapData.cpp
@@ -173,7 +173,7 @@ bool CMapData::LoadData(int id)
 				break;
 
 			case DP_COLA:
-				pObj->SetAlpha(stof(tmp));
+				pObj->SetAlpha(stoi(tmp));
 				break;
 
 			case DP_COLFLG:
